trunk/server/deploy_requirements.txt
django==1.4.1
South==0.7.6
django-tastypie==0.9.11
MySQL-python==1.2.3
django_nose==0.1.3
python-rest-client==0.3
httplib2==0.7.4
